Do not obfuscate but instead restart timer on startup for message by self
We fetch all ephemeral messages on startup and check if their deletion date is in the past, if so we delete them.
Until now we obfuscated all ephemeral messages sent by self if the application got killed and restarted, we now restart the timer if the delegation date is in the future and only obfuscate them when it is in the past (if the message is not be the self user we delete them when the deletion date passed and restart the timer if not).