Add v2 and v3 image asset proxies to declutter ZMAssetClientMessage
The `imageMessageData` accessor now returns different implementations of this protocol depending on the version of the underlying asset.
The two new added classes, `V2Asset` and `V3Asset` manage which cache and which cache key should be used to store the asset data and where to look for the image / file data etc.