remove unneccessary check
this check was preventing system messages being created for first time self member join events, b/c the brand new conversation has not had it’s conversation type set to group yet, and it defaults to invalid. Thus we were returning nil. However, in the first line of the method we already do event type checking, so most of these removed checks are unneccessary. Though, we assume that conversation on the BE is a group conversation.